+++ src/components/NativePicker/pickerStyles/index.android.js
